requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==4.0
gunicorn==20.1.0
djangorestframework==3.13.1
django-cors-headers==3.13.0